+++ glade_forms/lib/src/widgets/glade_form_builder.dart
@@ -9,7 +9,6 @@ class GladeFormBuilder<M extends GladeModel> extends StatelessWidget {
   // ignore: prefer-correct-callback-field-name, ok name
   final CreateModelFunction<M>? create;
   final M? value;
-  // ignore: prefer-correct-callback-field-name, ok name
   final GladeFormWidgetBuilder<M> builder;
   final Widget? child;